sam3d/requirements.p3d.txt
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@75ebeeaea0908c5527e7b1e305fbc7681382db47
flash_attn==2.8.3